Backend done
- simple backend for looking currency crypto price
