+++ backend/src/main.py
@@ -1,18 +1,6 @@
 from fastapi import FastAPI
-from src.http_client import CMCHTTPClient
-from src.config import settings
+from src.router import router as router_crypto
 
 app = FastAPI()
 
-cmc_client = CMCHTTPClient(
-    base_url="https://pro-api.coinmarketcap.com",
-    api_key=settings.CMC_API_KEY
-)
-
-@app.get("/cryptocurrencies")
-async def get_cryptocurrencies():
-    return await cmc_client.get_listing()
-
-@app.get("/cryptocurrencies/{currency_id}")
-async def get_cryptocurrency(currency_id: int):
-    return await cmc_client.get_currency(currency_id)
+app.include_router(router_crypto)
