+++ backend/src/router.py
@@ -1,0 +1,15 @@
+from fastapi import APIRouter
+from src.init import cmc_client
+
+router = APIRouter(
+    prefix = "/cryptocurrencies",
+    tags = ["Crypto Currencies"]
+)
+
+@router.get("")
+async def get_cryptocurrencies():
+    return await cmc_client.get_listing()
+
+@router.get("/{currency_id}")
+async def get_cryptocurrency(currency_id: int):
+    return await cmc_client.get_currency(currency_id)
